requirement.txt
###### Requirements without Version Specifiers ######
pillow
opencv-python
face_recognition
dlib
numpy
#
